+++ core/config/Abeille.config.php
@@ -15,19 +15,15 @@
     $abQueues["ctrlToCmd"] = array( "id" => 0x338, "max" => 2048 ); // Ctrl messages for AbeilleCmd
     $abQueues["parserToLQI"] = array( "id" => 0xE1, "max" => 2048 );
     $abQueues["parserToCli"] = array( "id" => 0x2D4, "max" => 1024 );
-    // $abQueues["parserToCmd"] = array( "id" => 0x3E6, "max" => 512 );
     $abQueues["parserToCmdAck"] = array( "id" => 0x3E7, "max" => 512 ); // Parser to cmd for 8000/8012/8702 statuses
-    $abQueues["parserToAbeille"] = array( "id" => 221, "max" => 512 ); // Parser Abeille, old path
-    $abQueues["parserToAbeille2"] = array( "id" => 222, "max" => 512 ); // Parser Abeille, new path
+    // $abQueues["parserToAbeille2"] = array( "id" => 222, "max" => 512 ); // Parser Abeille, new path
     $abQueues["xToCmd"] = array( "id" => 1212, "max" => 512 ); // AbeilleCmd inputs
     $abQueues["cmdToMon"] = array( "id" => 130, "max" => 512 ); // Messages to zigate (cmd to monitor)
     $abQueues["parserToMon"] = array( "id" => 131, "max" => 512 ); // Messages from zigate (parser to monitor)
     $abQueues["monToCmd"] = array( "id" => 132, "max" => 512 ); // Messages to cmd (addr update)
     $abQueues["parserToAssist"] = array( "id" => 141, "max" => 512 ); // Parser to EQ assistant
     $abQueues["assistToCmd"] = array( "id" => 142, "max" => 512 ); // Assistant to cmd
-    $abQueues["abeilleToAbeille"] = array( "id" => 121, "max" => 512 ); // ?
-    $abQueues["xmlToAbeille"] = array( "id" => 621, "max" => 512 ); // ?
-    $abQueues["cmdToAbeille"] = array( "id" => 321, "max" => 512 ); // ?
+    $abQueues["xToAbeille"] = array( "id" => 621, "max" => 512 ); // ?
     $GLOBALS['abQueues'] = $abQueues;
 
     // Tcharp38 note: WARNING: TOP PRIORITY is 1, not 5